feat: Auto reload on Devbox
This template does not auto reload. In Sandbox mode we have implemented so that `static` template auto reloads, but on Devbox we should not really hardcode that.

Sadly there is only the `live-server` package that makes this possible and for some reason it vies an EOL error when running on CSB... no idea what that is.

So the closest thing we get to this is using parcel.